Refactor to move App related stuff out of app
